Reader for sim profiles implemented (but not doing anything interesting) in the sm-dp-plus
