+++ sim-administration/simmanager/pom.xml
@@ -33,6 +33,11 @@
             <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
+            <artifactId>simcard-utils</artifactId>
+            <groupId>org.ostelco</groupId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>org.ostelco</groupId>
             <artifactId>es2plus4dropwizard</artifactId>
             <version>1.0-SNAPSHOT</version>
