+++ sim-administration/sm-dp-plus/pom.xml
@@ -29,6 +29,12 @@
         </dependency>
 
         <dependency>
+            <artifactId>simcard-utils</artifactId>
+            <groupId>org.ostelco</groupId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
             <groupId>org.ostelco</groupId>
             <artifactId>ostelco-dropwizard-utils</artifactId>
             <version>1.0-SNAPSHOT</version>
